feat: add ConfigResolversModule for improved dependency injection of config resolvers
