+++ src/quickapp/dial_core_services/dial_file_service.py
@@ -10,7 +10,9 @@ from injector import inject
 from pydantic import BaseModel, ConfigDict
 
 from quickapp.common.state_holder import StateHolder
-from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
+from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import (
+    FileLoadingSizeLimitResolver,
+)
 
 logger = logging.getLogger(__name__)
