+++ src/quickapp/mcp_tooling/_mcp_connection_manager.py
@@ -21,8 +21,8 @@ from quickapp.config.toolsets.authorization import (
     MCPApiKeyAuthorization,
 )
 from quickapp.config.toolsets.mcp import MCPProtocol, MCPServerInfo, MCPToolSet
-from quickapp.shared.config_resolvers.tool_timeout_resolver import ToolTimeoutResolver
 from quickapp.mcp_tooling._mcp_unauthorized_exception import MCPUnauthorizedException
+from quickapp.shared.config_resolvers.tool_timeout_resolver import ToolTimeoutResolver
 
 MAX_ITERATIONS = 1000
