+++ src/tests/unit_tests/dial_core_services_tests/test_dial_file_service.py
@@ -4,8 +4,10 @@ import pytest
 from aidial_client._exception import EtagMismatchError
 
 from quickapp.common.state_holder import StateHolder
-from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
 from quickapp.dial_core_services.dial_file_service import DialFileService
+from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import (
+    FileLoadingSizeLimitResolver,
+)
 
 DEFAULT_LIMIT = 10 * 1024 * 1024
