+++ src/tests/unit_tests/shared/test_external_url_fetcher.py
@@ -3,7 +3,9 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import httpx
 import pytest
 
-from quickapp.common.file_loading_size_limit_resolver import FileLoadingSizeLimitResolver
+from quickapp.shared.config_resolvers.file_loading_size_limit_resolver import (
+    FileLoadingSizeLimitResolver,
+)
 from quickapp.shared.external_fetch._external_fetch_settings import ExternalFetchSettings
 from quickapp.shared.external_fetch._external_url_fetch_policy_resolver import (
     ExternalUrlFetchPolicyResolver,
